+++ src/globus_sdk/transport/__init__.py
@@ -1,11 +1,21 @@
 from .encoders import FormRequestEncoder, JSONRequestEncoder, RequestEncoder
 from .requests import RequestsTransport
-from .retry import RetryCheckResult, RetryContext, RetryPolicy
+from .retry import (
+    RetryCheck,
+    RetryCheckFlags,
+    RetryCheckResult,
+    RetryCheckRunner,
+    RetryContext,
+    set_retry_check_flags,
+)
 
 __all__ = (
     "RequestsTransport",
-    "RetryPolicy",
+    "RetryCheck",
+    "RetryCheckFlags",
     "RetryCheckResult",
+    "RetryCheckRunner",
+    "set_retry_check_flags",
     "RetryContext",
     "RequestEncoder",
     "JSONRequestEncoder",