+++ tests/functional/auth/test_simple_auth_usage.py
@@ -84,9 +84,9 @@ def identities_multiple_response():
 
 
 @pytest.fixture
-def client(no_retry_policy):
+def client(no_retry_transport):
     class CustomAuthClient(globus_sdk.AuthClient):
-        retry_policy = no_retry_policy
+        transport_class = no_retry_transport
 
     return CustomAuthClient()